+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - Add ngOnChanges in the input observable of the table to be able to update the table when the input changes in the chips component
 - Task 1: add button to create a hero in a form into an expansion panel into a dialog and include it on top of the table
 - Bonus Task 3: Add options to delete and modify heroes
+- Bugfix: chips, if all the filters have been cleared returnt the full list